Code cleanup. Small changes to make the code more readable and clear. Unpacking of variables, etc. _convert_monet2mip_column_type and _convert_mip2monetdb_column_type should include only the MIP-types(real,int,text) so types 'clob' and 'bool' were deleted. `float` is not a MIP type, changed to `real`.
