+++ mipengine/common/resources/node_catalog.json
@@ -1,15 +1,15 @@
 {
   "globalNode": {
     "nodeId": "globalnode",
-    "rabbitmqURL": "127.0.0.1:5672",
-    "monetdbHostname": "192.168.63.129",
+    "rabbitmqURL": "192.168.124.129:5670",
+    "monetdbHostname": "192.168.124.129",
     "monetdbPort": "50000"
   },
   "localNodes": [
     {
       "nodeId": "localnode1",
-      "rabbitmqURL": "127.0.0.1:5673",
-      "monetdbHostname": "192.168.63.129",
+      "rabbitmqURL": "192.168.124.129:5671",
+      "monetdbHostname": "192.168.124.129",
       "monetdbPort": "50001",
       "data": {
         "pathologies": [
@@ -39,8 +39,8 @@
     },
     {
       "nodeId": "localnode2",
-      "rabbitmqURL": "127.0.0.1:5674",
-      "monetdbHostname": "192.168.63.129",
+      "rabbitmqURL": "192.168.124.129:5672",
+      "monetdbHostname": "192.168.124.129",
       "monetdbPort": "50002",
       "data": {
         "pathologies": [
